backend/requirements.txt
Flask
flask-cors
bcrypt
PyJWT
mysql-connector-python
pandas
numpy
networkx
scikit-learn
